askmynotes/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=4.0.0
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0
